pyproject.toml
[project]
name = "web-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.3",
    "modal>=1.2.4",
    "pillow>=12.0.0",
    "playwright>=1.56.0",
]
